refactor(workflow): consolidate workflow resource logic
